+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed 
+- Use workaround for IPv6 bug in TcpClient with Mono runtime used with Unity engine.
 
 ## [2.3.0] - 2019-09-02
 ### Added