+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.2.2-p1] - 2019-09-06
+### Changed 
+- Use workaround for IPv6 bug in TcpClient with Mono runtime used with Unity engine.
+
 ## [2.2.2] - 2019-07-02
 ### Changed
 - Don't synchronize the socket receive with the current thread context.